refactor: 로그인된 사용자와 로그인 되지 않은 사용자 분리
